task_1/requirements.txt
psycopg2-binary==2.9.7
Faker==18.9.0